Input file name is a command-line argument
Used to be hard-coded for Tom's machine. Now a command-line argument
-f <filename>
(eg, for s3d input dataset, Nek5000, etc.)

Unit tests also have the added argument. The value of the argument in
the unit tests is the cmake variable ${s3d_infile}. The value for this
variable can be set during the cmake configuration as
-Ds3d_infile=/path/to/s3d/file