Decoder::VolPt_tmesh works for first level of tmesh
VolPt_tmesh takes a parameter value to decode, computes the anchors,
local knot vectors, and calls a new version of VolPt.
Added a new version of Decoder::VolPt that takes precomputed basis
function (computed by OneBasisFun given a local knot vector). This
version of VolPt also takes custom set of control points and weights.

Tested and seems correct for geometry decoding. Does not work yet for
adaptive decoding because the first adaptation causes knot insertion
(not implemented yet) to be required across the boundaries of tensor
products. This is the next step