Fixed bug in tmesh decoding
Tmesh::anchors() changed how anchors are computed.
Decoder::VolPt_tmesh() changed how anchors are used.

Anchors now start counting at 0. They are first computed for the
purpose of forming the basis functions. For this, they are essentially
the same as the result of FindSpan, the span containing the parameter
value, but offset by -p so that they start at 0. This is the same
as the index of the corresponding basis function.

Then, when anchors are used as the center of the local knot vector, they
are offset by (p+1)/2 so that they are the center of the local knot
vector. The knot vector is found by going left and right from the
center, looking for knot line intersections in the tmesh.

This works now for any degree when there is only one tensor product (ie,
fixed example). However, I still need
to write a knot_intersections function similar, but slightly different,
from the knot_intersections used to find local knot vectors. This new
knot intersections function needs to be called from anchors().
This is the next step.