Row major order for Eigen matrix and array
Changed in "using" statements in mfa.hpp, storing Eigen matrices and
arrays in row-major order. Impact on performance to be determined. In
principle, most accesses should be per row, so row-major ought to be
appropriate.

Also removed duplicate "using" statements in block_base.hpp and
block.hpp, so that mfa.hpp is the only place this needs to be changed.

Row-major storage is also needed for correct mapping of Eigen matrices to SYCL
buffers and accessors for GPU kernels that are written to process one
row of the array independently from the others in row-major order.