WIP: debugging tom-tspline branch
- Encoder::NewKnots_full calls NewKnots::TempFirstErrorSpan instead of
FirstErrorSpan until the adaptive encode is implemented. Encodes a full
tensor product of control points in an explicit matrix outside of the
tmesh. Calls a new versionn of Decoder::VolPt that takes the control
points as inputs.
- Once the adaptive encoding is in place, call FirstErrorSpan and
deprecate/remove TempFirstErrorSpan

running examples:

./adaptive -d 2 -m 1 -q 2 -u 1 -v 4 -n 5 -e 1e0 (no levels of refinement, only base level)
runs but answers are wrong (decoding 0 and nan)

./adaptive -d 2 -m 1 -q 2 -u 1 -v 4 -n 5 -e 1e-1 (1 level of refinement)
crashes