WIP: debugging tom-tspline
Fixed a bug in Encoder::Encode(). When calling Encode multiple times, as in
adaptive encoding, MFA_Data::N (basis functions) needs to be reset to the
zero matrix. I was resizing it in case the number of control points
changed, but this does not set to zero. Instead of resize, assign to
zero matrix of the correct new size (which resizes as a side effect).