Renamed Tmesh::new_max_side()
To new_side() to reflect that the new side can be either on the minimum
or the maximum side of the existing tensor. Also cleaned up that code a
little more.

Remaining is still to change split_ctrl_pts to work with either a new
minimum or a new maximum side. Currently correct only if the new side is
on the maximum side.