Use local copy of opts.h for examples and tests
Put a local copy of opts.h into the examples directory of mfa. Not using
diy's copy of opts.h, which is in diy's examples directory, because it's
possible to install diy without examples (this is the default spack
installation), in which case there is no opts.h in diy.