Added parent_tensor_idx to NewKnots::InsertKnots
The parent tensor is needed in order to pass to
MFA_Data::KnotInsertion().

The global solve version of InsertKnots forms a union of new and old
knots, and also finds the parent_tensor_idx of the first knot being
inserted (for single knot insertion).

The local solve version of InsertKnots calls that global solve version
to do the set union and to get the parent_tensor_idx. It then passes the
parent_tensor to MFA_Data::KnotInsertion().