Fixed VTK converter writing control points
Correctly goes through all tensor products in the tmesh now, for both
geometry and science variables (previously was hard-coded for only first
tensor).

NB: the order of tensors in the tmesh, and hence the ordering of the control
points in the VTK file, is scattered. Use only points mode (not surface
or outline or mesh) in ParaView or VisIt to view the control points.