Added versions of knot insertion for existing knot
In the current implementation, the new knot is inserted into the t-mesh
before calling MFA_Data::KnotInsertion. This means that the new knots do
not need to be created or passed. The knot insertion functions in
mfa_data.hpp (KnotInsertion, VolKnotIns, CurveKnotIns) were modified by
making two versions of each (New... and Exist...) depending on whether the
new knot was not yet inserted or already inserted.