Fixed bug in local solve
Optimization now converges correctly to low error, with reasonable
control point locations that are close to the inital guess from Boehm
knot insertion. Problem was the least squares error function was
comparing the decoded point with the wrong coordinate of the input
domain point (geometry coordinate instead of science value), which was
driving the optimizer away from the correct solution.