Fixed bug in Encoder::LocalSolveCtrlPts()
Setting constraints incorrectly in higher dimensions; simple typo in
indexing that is fixed now.

Tested doing one refinement in 2-d, works correctly now. The main
problem is the high number of iterations needed to converge, and the
fact that each iteration is expensive because VolPt_tmesh is not
optimized yet.