Fixed a bug in Tmesh::scatter_ctrl_pts()
In 2d, when degree is odd (p=3), wrong numbers of pointers were being
distributed to the tensor products, causing an assertion to fail. This
is fixed now. The problem was with neighbor tensors, needed to check whether
point is in the neighbor and whether its level is less than or equal to
the neighbor tensor.