Code cleanups and testing
Decided that there was no bug in Tmesh::scatter_ctrl_pts(). Differences
in results between global and local solve versions can be explained
because control points in a single global tensor solution do not have an
exact mapping to control points in the local individual tensors of a
t-mesh. In other words, scattering control points from a global solution
may be close, but technically it is not correct, and so we shouldn't
expect the error to match. Indeed, for the t-mesh, the error is better
for the local solve than the global solve.