Fixed bug in decoding with higher number of rounds
Tmesh::knot_intersections() was reworked to fix some logic bugs. As part
of this fix, an argument was added to Tmesh::in() and the functions that
call it to specify whether 0.5 should be added to the target point (for
even degree control points). Added Tmesh::in_dim(), similar to in(), but
for a single dimension.