Fixed Tmesh::subset_ctrl_pts()
The different cases for computing sub_starts were incorrect in the
previous commit. The new tensor does not need to be only to one side or
the other other parent; it can also be in the middle of the parent. It
is true that when a new tensor is split off from a parent, it is to one
side, but only in the dimension being split. In the other dimensions
(eg, in 2d), the child can be in the interior of the parent.