Fixed bug in tspline
In a particular example with p=3, the tensor to be appended intersected
several other existing tensors, covering one completely. Added logic in
Tmesh::append_tensor() to delete an existing tensor in this case.

Also added Tmesh::delete_pointer() and change_pointer() to support
deleting a tensor and renumbering the prev/next pointers of its
neighbors.