Adds FastVolPt, an optimized decoding method for 1D MFAs
This function can only be used with no weights, so MFA_NO_WEIGHTS
must be defined. It does not support TMesh.

This commit also adds two support structs, BasisFunInfo and
FastDecodeInfo, which primarily contain buffers which are used
by FastVolPt.

FastVolPt also calls a new function for computing basis functions,
FastBasisFuns. This method is the same as OrigBasisFuns but
does not define Eigen objects, which speeds up execution.

Finally, a few changes were made to Decoder. Some information is
precomputed during the construction of a Decoder (like q, the
degree + 1, or ds, the stride in the control point subvolume).
Precomputing this once allows us to streamline FastVolPt.