Fixed Tmesh::in()
Checks correctly whether a point is in the tensor, finding the idx of
the weights in order to check for MFA_NAW. This was hacked up before,
and wasn't correct all the time. Should be robust now. Also added a
version of Tmesh::in() that doesn't do these checks when not desired.

Unified dimension local encoding works again (was crashing previously).
Generates some blocky artifacts that still need to be debugged.

Separable dimenson local encoding currently broken, still needs to be
restored to working order and then the result needs to be checked.