Adding new side tensor and splitting control pts
Fixed a bug where the number of knots and number of control points
disagreed when appending a tensor -> intersecting with existing tensors
-> adding a new side tensor -> splitting control points between existing
and new side tensor.

The order of computing the local control point index of the split was
incorrect because it was being computed in split_ctrl_pts() after the knot
min/maxs were adjusted in new_side(). Instead compute the split control point
index of the existing tensor in new_side() before adjusting knot
mins/maxs and pass the control point index to split_ctrl_pts().