Fixed bug in AddNewTensors
It was possible for a tensor to be added to intersect an existing tensor
within a small distance, leaving insufficient control points. This
happens when there are multiple tensor additions in one iteration, and
the existing tensors get split into smaller ones. This case cannot be
checked earlier during Refine.

Solution is to invalidate this tensor to be added, mark the intersecting
existing tensors as not done, do not increase the parent level, and let
the existing tensors be rechecked in the next iteration.