Simplifies Bbox to not account for translation when mapping between coordinate systems
Before, the Bbox change of coordinates would map the minimum corner to the origin in rotated space.
This was done by manually translating all elements by the minimum corner.
However, this meant that two bounding boxes with the same orientation but different minimal corners would map different points to the origin.

After this commit, bounding boxes with the same orientation always map the same points in cartesian space to the same points in rotated space, regardless of their corners.